+++ Tests/OpenAITests/StreamingSessionIntegrationTests.swift
@@ -103,26 +103,4 @@ final class StreamingSessionIntegrationTests: XCTestCase {
     }
 }
 
-struct WorkSimulatingMockMiddleware: OpenAIMiddleware {
-    func intercept(request: URLRequest) -> URLRequest {
-        simulateBusyThread(duration: 0.1)
-        return request
-    }
-    
-    func interceptStreamingData(request: URLRequest?, _ data: Data) -> Data {
-        simulateBusyThread(duration: 0.1)
-        return data
-    }
-    
-    func intercept(response: URLResponse?, request: URLRequest, data: Data?) -> (response: URLResponse?, data: Data?) {
-        simulateBusyThread(duration: 0.1)
-        return (response, data)
-    }
-    
-    private func simulateBusyThread(duration: TimeInterval) {
-        let end = Date().addingTimeInterval(duration)
-        while Date() < end {
-            _ = UUID().uuidString.hashValue  // Some pointless work
-        }
-    }
-}
+
